@@ -1,35 +1,46 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'dart:convert';
 import 'package:flutter_picker/flutter_picker.dart';
 import 'PickerData.dart';
-//import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 
-void main() => runApp(new MyApp());
+void main() => runApp(MyApp());
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => new _MyAppState();
+  _MyAppState createState() => _MyAppState();
 }
 
+final String _fontFamily = Platform.isWindows ? "Roboto" : "";
+
 class _MyAppState extends State<MyApp> {
+
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
+    return MaterialApp(
         theme: ThemeData(
           primarySwatch: Colors.blue,
+          fontFamily: _fontFamily,
+          primaryTextTheme: TextTheme().apply(fontFamily: _fontFamily),
         ),
 
-//        localizationsDelegates: [
-//          PickerLocalizationsDelegate.delegate, // 如果要使用本地化，请添加此行，则可以显示中文按钮
-//          GlobalMaterialLocalizations.delegate,
-//          GlobalWidgetsLocalizations.delegate,
-//        ],
-//        supportedLocales: [
-//          const Locale('en', 'US'),
-//          const Locale('zh', 'CH'),
-//        ],
+        localizationsDelegates: [
+          PickerLocalizationsDelegate.delegate, // 如果要使用本地化，请添加此行，则可以显示中文按钮
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+        ],
+        supportedLocales: [
+          const Locale('en', 'US'),
+          const Locale('zh', 'CH'),
+          const Locale('ko', 'KO'),
+          const Locale('it', 'IT'),
+          const Locale('ar', 'AR'),
+          const Locale('tr','TR')
+        ],
 
-        home: new MyHomePage());
+        home: MyHomePage());
   }
 }
 
@@ -41,14 +52,14 @@ class MyHomePage extends StatefulWidget {
 class _MyHomePageState extends State<MyHomePage> {
   final double listSpec = 4.0;
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-  String stateText;
+  String stateText = "";
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       key: _scaffoldKey,
       appBar: AppBar(
-        title: Text('Picker'),
+        title: Text('Picker ${stateText.isEmpty ? "" : " - " + stateText}'),
         automaticallyImplyLeading: false,
         elevation: 0.0,
       ),
@@ -57,80 +68,75 @@ class _MyHomePageState extends State<MyHomePage> {
         alignment: Alignment.topCenter,
         child: ListView(
           children: <Widget>[
-            (stateText != null) ? Text(stateText) : Container(),
-            RaisedButton(
-              child: Text('Picker Show'),
-              onPressed: () {
+            ListTile(
+              title: Text('1. Picker Show'),
+              onTap: () {
                 showPicker(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Modal'),
-              onPressed: () {
+            ListTile(
+              title: Text('2. Picker Show Modal'),
+              onTap: () {
                 showPickerModal(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Icons'),
-              onPressed: () {
+            ListTile(
+              title: Text('3. Picker Show Icons'),
+              onTap: () {
                 showPickerIcons(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show (Array)'),
-              onPressed: () {
+            ListTile(
+              title: Text('4. Picker Show (Array)'),
+              onTap: () {
                 showPickerArray(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Number'),
-              onPressed: () {
+            ListTile(
+              title: Text('5. Picker Show Number'),
+              onTap: () {
                 showPickerNumber(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Number FormatValue'),
-              onPressed: () {
+            ListTile(
+              title: Text('6. Picker Show Number FormatValue'),
+              onTap: () {
                 showPickerNumberFormatValue(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Date'),
-              onPressed: () {
+            ListTile(
+              title: Text('7. Picker Show Date'),
+              onTap: () {
                 showPickerDate(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Datetime'),
-              onPressed: () {
+            ListTile(
+              title: Text('8. Picker Show Datetime'),
+              onTap: () {
                 showPickerDateTime(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Date (Custom)'),
-              onPressed: () {
+            ListTile(
+              title: Text('9. Picker Show Date (Custom)'),
+              onTap: () {
                 showPickerDateCustom(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Datetime (24)'),
-              onPressed: () {
+            ListTile(
+              title: Text('10. Picker Show Datetime (24)'),
+              onTap: () {
                 showPickerDateTime24(context);
               },
             ),
-            SizedBox(height: listSpec),
-            RaisedButton(
-              child: Text('Picker Show Date Range'),
-              onPressed: () {
+            ListTile(
+              title: Text('11. Picker Show Datetime (Round background)'),
+              onTap: () {
+                showPickerDateTimeRoundBg(context);
+              },
+            ),
+            ListTile(
+              title: Text('12. Picker Show Date Range'),
+              onTap: () {
                 showPickerDateRange(context);
               },
             ),
@@ -143,9 +149,9 @@ class _MyHomePageState extends State<MyHomePage> {
   showPicker(BuildContext context) {
     Picker picker = Picker(
       adapter: PickerDataAdapter<String>(pickerdata: JsonDecoder().convert(PickerData)),
-      changeToFirst: true,
+      changeToFirst: false,
       textAlign: TextAlign.left,
-      textStyle: const TextStyle(color: Colors.blue),
+      textStyle: TextStyle(color: Colors.blue, fontFamily: _fontFamily),
       selectedTextStyle: TextStyle(color: Colors.red),
       columnPadding: const EdgeInsets.all(8.0),
       onConfirm: (Picker picker, List value) {
@@ -156,17 +162,34 @@ class _MyHomePageState extends State<MyHomePage> {
     picker.show(_scaffoldKey.currentState);
   }
 
-  showPickerModal(BuildContext context) {
-    Picker(
+  showPickerModal(BuildContext context) async {
+    final result = await Picker(
       adapter: PickerDataAdapter<String>(pickerdata: JsonDecoder().convert(PickerData)),
       changeToFirst: true,
       hideHeader: false,
       selectedTextStyle: TextStyle(color: Colors.blue),
-      onConfirm: (Picker picker, List value) {
+      // builderHeader: (context) {
+      //   final picker = PickerWidget.of(context);
+      //   return picker?.data?.title ?? Text("xxx");
+      // },
+      onConfirm: (picker, value) {
         print(value.toString());
         print(picker.adapter.text);
+        Navigator.push(context, MaterialPageRoute(builder: (context) => Scaffold(
+          appBar: AppBar(title: Text("Hello")),
+          body: Center(child: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text("You click the Confirm button."),
+              SizedBox(height: 32),
+              Text("result: \n ${picker.adapter.text}")
+            ],
+          )),
+        )));
       }
-    ).showModal(this.context); //_scaffoldKey.currentState);
+    ).showModal(this.context); //_sca
+    print("result: $result");// ffoldKey.currentState);
   }
 
   showPickerIcons(BuildContext context) {
@@ -176,7 +199,11 @@ class _MyHomePageState extends State<MyHomePage> {
             PickerItem(text: Icon(Icons.more)),
             PickerItem(text: Icon(Icons.aspect_ratio)),
             PickerItem(text: Icon(Icons.android)),
-            PickerItem(text: Icon(Icons.menu)),
+            PickerItem(text: Icon(Icons.menu), children: [
+              // 测试：多加了一维数据
+              PickerItem(text: Icon(Icons.account_box)),
+              PickerItem(text: Icon(Icons.analytics)),
+            ]),
           ]),
           PickerItem(text: Icon(Icons.title), value: Icons.title, children: [
             PickerItem(text: Icon(Icons.more_vert)),
@@ -200,7 +227,7 @@ class _MyHomePageState extends State<MyHomePage> {
           PickerItem(text: Icon(Icons.close), value: Icons.close),
         ]),
         title: Text("Select Icon"),
-        selectedTextStyle: TextStyle(color: Colors.blue),
+        selectedTextStyle: TextStyle(color: Colors.blue, fontSize: 12),
         onConfirm: (Picker picker, List value) {
           print(value.toString());
           print(picker.getSelectedValues());
@@ -213,7 +240,7 @@ class _MyHomePageState extends State<MyHomePage> {
     Picker(
         adapter: PickerDataAdapter<String>(pickerdata: JsonDecoder().convert(PickerData)),
         hideHeader: true,
-        title: new Text("Select Data"),
+        title: Text("Select Data"),
         selectedTextStyle: TextStyle(color: Colors.blue),
         onConfirm: (Picker picker, List value) {
           print(value.toString());
@@ -225,14 +252,14 @@ class _MyHomePageState extends State<MyHomePage> {
   showPickerArray(BuildContext context) {
     Picker(
         adapter: PickerDataAdapter<String>(
-            pickerdata: JsonDecoder().convert(PickerData2),
-            isArray: true,
+          pickerdata: JsonDecoder().convert(PickerData2),
+          isArray: true,
         ),
         hideHeader: true,
         selecteds: [3, 0, 2],
         title: Text("Please Select"),
         selectedTextStyle: TextStyle(color: Colors.blue),
-        cancel: FlatButton(onPressed: () {
+        cancel: TextButton(onPressed: () {
           Navigator.pop(context);
         }, child: Icon(Icons.child_care)),
         onConfirm: (Picker picker, List value) {
@@ -307,12 +334,12 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   showPickerDateCustom(BuildContext context) {
-    new Picker(
+    Picker(
         hideHeader: true,
-        adapter: new DateTimePickerAdapter(
+        adapter: DateTimePickerAdapter(
           customColumnType: [2,1,0,3,4],
         ),
-        title: new Text("Select Data"),
+        title: Text("Select Data"),
         selectedTextStyle: TextStyle(color: Colors.blue),
         onConfirm: (Picker picker, List value) {
           print((picker.adapter as DateTimePickerAdapter).value);
@@ -321,8 +348,8 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   showPickerDateTime(BuildContext context) {
-    new Picker(
-        adapter: new DateTimePickerAdapter(
+    Picker(
+        adapter: DateTimePickerAdapter(
           type: PickerDateTimeType.kYMD_AP_HM,
           isNumberMonth: true,
           //strAMPM: const["上午", "下午"],
@@ -330,9 +357,12 @@ class _MyHomePageState extends State<MyHomePage> {
           monthSuffix: "月",
           daySuffix: "日",
           minValue: DateTime.now(),
+          minuteInterval: 30,
+          minHour: 1,
+          maxHour: 23,
           // twoDigitYear: true,
         ),
-        title: new Text("Select DateTime"),
+        title: Text("Select DateTime"),
         textAlign: TextAlign.right,
         selectedTextStyle: TextStyle(color: Colors.blue),
         delimiter: [
@@ -362,41 +392,41 @@ class _MyHomePageState extends State<MyHomePage> {
   showPickerDateRange(BuildContext context) {
     print("canceltext: ${PickerLocalizations.of(context).cancelText}");
 
-    Picker ps = new Picker(
+    Picker ps = Picker(
         hideHeader: true,
-        adapter: new DateTimePickerAdapter(type: PickerDateTimeType.kYMD, isNumberMonth: true),
+        adapter: DateTimePickerAdapter(type: PickerDateTimeType.kYMD, isNumberMonth: true),
         onConfirm: (Picker picker, List value) {
           print((picker.adapter as DateTimePickerAdapter).value);
         }
     );
 
-    Picker pe = new Picker(
+    Picker pe = Picker(
         hideHeader: true,
-        adapter: new DateTimePickerAdapter(type: PickerDateTimeType.kYMD),
+        adapter: DateTimePickerAdapter(type: PickerDateTimeType.kYMD),
         onConfirm: (Picker picker, List value) {
           print((picker.adapter as DateTimePickerAdapter).value);
         }
     );
 
     List<Widget> actions = [
-      FlatButton(
+      TextButton(
           onPressed: () {
             Navigator.pop(context);
           },
-          child: new Text(PickerLocalizations.of(context).cancelText)),
-      FlatButton(
+          child: Text(PickerLocalizations.of(context).cancelText)),
+      TextButton(
           onPressed: () {
             Navigator.pop(context);
             ps.onConfirm(ps, ps.selecteds);
             pe.onConfirm(pe, pe.selecteds);
           },
-          child: new Text(PickerLocalizations.of(context).confirmText))
+          child: Text(PickerLocalizations.of(context).confirmText))
     ];
 
     showDialog(
         context: context,
         builder: (BuildContext context) {
-          return new AlertDialog(
+          return AlertDialog(
             title: Text("Select Date Range"),
             actions: actions,
             content: Container(
@@ -416,15 +446,19 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   showPickerDateTime24(BuildContext context) {
-    new Picker(
-        adapter: new DateTimePickerAdapter(
+    Picker(
+        adapter: DateTimePickerAdapter(
             type: PickerDateTimeType.kMDYHM,
             isNumberMonth: true,
             yearSuffix: "年",
             monthSuffix: "月",
-            daySuffix: "日"
+            daySuffix: "日",
+            minHour: 8,
+            maxHour: 19,
+            yearBegin: 1950,
+            yearEnd: 1998,
         ),
-        title: new Text("Select DateTime"),
+        title: Text("Select DateTime"),
         onConfirm: (Picker picker, List value) {
           print(picker.adapter.text);
         },
@@ -434,6 +468,55 @@ class _MyHomePageState extends State<MyHomePage> {
           });
         }
     ).show(_scaffoldKey.currentState);
+  }
+
+  /// 圆角背景
+  showPickerDateTimeRoundBg(BuildContext context) {
+    var picker = Picker(
+        backgroundColor: Colors.transparent,
+        headerDecoration: BoxDecoration(
+            border: Border(bottom: BorderSide(color: Colors.black12, width: 0.5))
+        ),
+        adapter: DateTimePickerAdapter(
+            type: PickerDateTimeType.kMDYHM,
+            isNumberMonth: true,
+            yearSuffix: "年",
+            monthSuffix: "月",
+            daySuffix: "日"
+        ),
+        delimiter: [
+          PickerDelimiter(column: 3, child: Container(
+            width: 8.0,
+            alignment: Alignment.center,
+          )),
+          PickerDelimiter(column: 5, child: Container(
+            width: 12.0,
+            alignment: Alignment.center,
+            child: Text(':', style: TextStyle(fontWeight: FontWeight.bold)),
+            color: Colors.white,
+          )),
+        ],
+        title: Text("Select DateTime"),
+        onConfirm: (Picker picker, List value) {
+          print(picker.adapter.text);
+        },
+        onSelect: (Picker picker, int index, List<int> selecteds) {
+          this.setState(() {
+            stateText = picker.adapter.toString();
+          });
+        }
+    );
+
+    showModalBottomSheet(context: context, backgroundColor: Colors.transparent, builder: (context) {
+      return Material(
+        color: Colors.white,
+        borderRadius: BorderRadius.only(topLeft: Radius.circular(10), topRight: Radius.circular(10)),
+        child: Container(
+          padding: const EdgeInsets.only(top: 4),
+          child: picker.makePicker(null, true),
+        )
+      );
+    });
   }
 
 }
