@@ -15,7 +15,10 @@ const PickerData = '''
                     5,
                     6,
                     7,
-                    8
+                    8,
+                    555,
+                    666,
+                    999
                 ]
             },
             {
@@ -43,15 +46,26 @@ const PickerData = '''
                     55,
                     66,
                     77,
-                    88
+                    88,
+                    99,
+                    1010,
+                    1111,
+                    1212,
+                    1313,
+                    1414,
+                    1515,
+                    1616
                 ]
             },
             {
                 "b3": [
-                    99,
                     1010,
                     1111,
-                    1212
+                    1212,
+                    1313,
+                    1414,
+                    1515,
+                    1616
                 ]
             }
         ]
@@ -77,6 +91,14 @@ const PickerData = '''
                     "aaa",
                     "bbb",
                     "ccc"
+                ]
+            },
+            {
+                "c4": [
+                    "a1",
+                    "b1",
+                    "c1",
+                    "d1"
                 ]
             }
         ]
